+++ sourceCode/gui/dialogs/qdlgpathshaping.cpp
@@ -175,7 +175,7 @@ void CQDlgPathShaping::on_qqMaxLength_editingFinished()
     IF_UI_EVENT_CAN_READ_DATA
     {
         bool ok;
-        double newVal=ui->qqMaxLength->text().toFloat(&ok);
+        double newVal=ui->qqMaxLength->text().toDouble(&ok);
         if (ok)
         {
             App::appendSimulationThreadCommand(SET_ELEMENTMAXLENGTH_PATH_OLD_SHAPINGGUITRIGGEREDCMD,App::currentWorld->sceneObjects->getLastSelectionHandle(),-1,newVal);
@@ -215,7 +215,7 @@ void CQDlgPathShaping::on_qqScalingFactor_editingFinished()
     IF_UI_EVENT_CAN_READ_DATA
     {
         bool ok;
-        double newVal=ui->qqScalingFactor->text().toFloat(&ok);
+        double newVal=ui->qqScalingFactor->text().toDouble(&ok);
         if (ok)
         {
             App::appendSimulationThreadCommand(SET_SCALINGFACTOR_PATH_OLD_SHAPINGGUITRIGGEREDCMD,App::currentWorld->sceneObjects->getLastSelectionHandle(),-1,newVal);
