+++ sourceCode/gui/dialogs/qdlgtextures.cpp
@@ -308,7 +308,7 @@ void CQDlgTextures::_setTextureConfig(int index)
 {
     QLineEdit* ww[6]={ui->qqX,ui->qqY,ui->qqZ,ui->qqAlpha,ui->qqBeta,ui->qqGamma};
     bool ok;
-    double newVal=ww[index]->text().toFloat(&ok);
+    double newVal=ww[index]->text().toDouble(&ok);
     if (ok)
     {
         SSimulationThreadCommand cmd;
@@ -330,7 +330,7 @@ void CQDlgTextures::_setTextureScaling(int index)
 {
     QLineEdit* ww[2]={ui->qqU,ui->qqV};
     bool ok;
-    double newVal=ww[index]->text().toFloat(&ok);
+    double newVal=ww[index]->text().toDouble(&ok);
     if (ok)
     {
         SSimulationThreadCommand cmd;
